feat(api-profiles): add MaybeAuthenticatedUser request extractor
This implied quite a refactor of the authentication middleware but
it is way better now. 401 Unauthorized errors are now thrown by the
request extractors and not by the middleware itself.

This also permits using the authentication middleware on the project
level (wrapped in the startup.rs file) instead of specifying it on every
request handler.